z09liner/requirements.txt
asgiref==3.3.4
Django==3.2.4
pytz==2021.1
sqlparse==0.4.1
requests==2.25.1
Pillow==8.2.0
django-mptt~=0.12.0
django-filter~=2.4.0
djangorestframework~=3.12.4
channels~=3.0.3
djangorestframework-recursive~=0.1.2
docker~=5.0.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
django-ckeditor==6.1.0
django-js-asset==1.2.2
idna==2.10
urllib3==1.26.5
django-summernote==0.8.11.6
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.5
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
hyperlink==21.0.0
incremental==21.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.2.1
Twisted==21.2.0
txaio==21.2.1
websocket-client==1.1.0
zope.interface==5.4.0
